some fixes from homework with fixtures, new tests added

+++ niffler-e-2-e-tests-python/databases/spend_db.py
@@ -33,5 +33,6 @@ class SpendDb:
     def delete_category(self, category_id: str):
         with Session(self.engine) as session:
             category = session.get(Category, category_id)
-            session.delete(category)
-            session.commit()
+            if category:
+                session.delete(category)
+                session.commit()
